+++ CHANGES.rst
@@ -13,6 +13,8 @@ Unreleased
 
 * Fix bug in :meth:`~Reader.enable_search()` that caused it to fail
   if search was already enabled and the reader had any entries.
+* Add an ``entry`` argument to :meth:`~Reader.get_entries`,
+  for symmetry with :meth:`~Reader.search_entries`.
 
 
 Version 0.19